show detail todo by id
